requirements.txt
coreferee
spacy
llama-index-graph-stores-neo4j
numpy
pip
transformers
torch
# langchain-neo4j
# langchain-openai
